pyproject.toml
[project]
name = "vectorless-rag-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111.0,<0.116",
    "httpx>=0.27.0,<0.29",
    "langchain>=0.2.0,<0.4",
    "langchain-community>=0.2.19",
    "langchain-google-genai>=1.0.0,<3.0",
    "langgraph>=0.1.0,<0.3",
    "llama-index-core>=0.10.0,<0.12",
    "llama-index-llms-gemini>=0.1.0,<0.4",
    "llama-index-readers-file>=0.1.0,<0.4",
    "pydantic>=2.7.0,<3.0",
    "pydantic-settings>=2.2.0,<3.0",
    "pypdf>=5.9.0",
    "pytest>=8.2.0,<9.0",
    "pytest-asyncio>=0.23.0,<0.26",
    "python-dotenv>=1.0.0,<2.0",
    "python-multipart>=0.0.9,<0.1",
    "pyvis>=0.3.2",
    "streamlit>=1.35.0,<2.0",
    "uvicorn[standard]>=0.29.0,<0.35",
]
